+++ s/loaders/index.ts
@@ -1,7 +1,3 @@
 
-export * from "./parts/ascii-anim.js"
-export * from "./parts/error-display.js"
-export * from "./make-loader.js"
-export * from "./mock-loader.js"
-export * from "./types.js"
+export * as loaders from "./index.barrel.js"