+++ s/loaders/make.ts
@@ -4,9 +4,7 @@ import {earth} from "./parts/anims.js"
 import type {Content} from "../ui/types.js"
 import {ErrorDisplay} from "./parts/error-display.js"
 
-export * as anims from "./parts/anims.js"
-
-export function makeLoader(
+export function make(
 		loading: () => Content = earth,
 		error: (error: any) => Content = (error: any) => ErrorDisplay(error),
 	): Loader {