+++ s/loaders/mock.ts
@@ -1,7 +1,7 @@
 
 import {Loader} from "./types.js"
 
-export function mockLoader(): Loader {
+export function mock(): Loader {
 	return (op, ready) => op.select({
 		ready,
 		loading: () => `[loading]`,